refactor: split dashboard context
