+++ src/features/events/components/EventsList/index.tsx
@@ -4,16 +4,17 @@ import { EventItem } from './parts/EventItem'
 import { NavigationFilter } from './parts/NavigationFilter'
 import { NavigationView } from './parts/NavigationView'
 import { Nav, List, ListItem } from './styled'
-import { ViewType } from './types'
 
-import { useDashboardContext } from '../../contexts/dashboard'
+import { useEventFilterContext } from '../../contexts/event-filter'
+import { useEventViewContext, ViewType } from '../../contexts/event-view'
 import { useEvents } from '../../hooks/useEvents'
 
 /**
  * Renders a list of events, with filtering/sorting/view type options.
  */
 export const EventsList: FC = () => {
-  const { filter, setFilter, view, setView } = useDashboardContext()
+  const { filter, setFilter } = useEventFilterContext()
+  const { view, setView } = useEventViewContext()
   const { events, isLoading, error } = useEvents(filter)
 
   // Avoid rendering UI if event loading failed.
