--- src/features/events/components/EventsList/types.ts
@@ -1,4 +1,0 @@
-export enum ViewType {
-  GRID = 'GRID',
-  LIST = 'LIST',
-}
